requirements.txt
amqp==5.3.1
asgiref==3.8.1
awsebcli==3.21.0
billiard==4.2.1
botocore==1.35.78
celery==5.4.0
cement==2.10.14
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==4.2.6
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-timezone-field==7.0
djangorestframework==3.15.2
idna==3.10
importlib_metadata==8.5.0
jmespath==1.0.1
kombu==5.4.2
mysqlclient==2.2.6
pathspec==0.10.1
prompt_toolkit==3.0.48
PyMySQL==1.1.1
pypiwin32==223
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
semantic-version==2.10.0
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.3
termcolor==2.5.0
tzdata==2024.2
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
zipp==3.21.0